In smb2.lease, tolerate Likewise behavior.
Modify smb2.lease.upgrade3 and smb2.lease.break to accept Likewise's
auto fallback on attempts to upgrade a lease. This results in the
client getting an 'RH' lease in cases where Windows and Samba will
only return an 'R' lease.

Modify several tests to wait up to 500ms for an interim response to
be sent from the server. Likewise defers the interim response with
the expectation that the break will clear and the create succeed
without the need to send an interim response at all.

Modify smb2.lease.breaking3 to accept a direct break to 'None' instead
of to 'R' and finally 'None'. The direct break to 'None' is the behavior
specified by MS-FSA.