Correct smb2.notify.rec torture test.
Add handle closes required to avoid hangs on Win2k12 or Likewise.

Remove forced failure of this test case. Likewise is returning 9
events as was previously expected.

Change a torture_warning to a torture_comment. Likewise is following
the spec, but the original test was not. The old comment implied that
Likewise was doing something wrong.